requirements.txt
fastapi
uvicorn
pika
python-multipart
pillow
